REFACTOR: functions to get wpi of city moved to weatherpleasantry.go; search single city webpage functionality moved into its own file
