+++ main_test.go
@@ -1,11 +1,11 @@
 package main
 
 import (
+	"github.com/Tris20/FairFareFinder/src/go_files/weather_pleasantness"
 	"io/ioutil"
 	"os"
 	"testing"
 	"time"
-  "github.com/Tris20/FairFareFinder/src/go_files/weather_pleasantness"
 )
 
 // test for displayForecastData function
