+++ src/go_files/json_functions/README.md
@@ -1,0 +1,1 @@
+NOTE this dir will be removed when we move to SQL. it's just to keep existing functionality till that change has happened
